% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Clean data}
\usage{
cleanData(
  dataName,
  includeIDs = NULL,
  excludeIDs = NULL,
  removeMultiallelicSites = FALSE,
  remove.nocall.pct = 0.2,
  QD.threshold = 5,
  saveRData = FALSE,
  savePath = NULL,
  label = NULL
)
}
\arguments{
\item{dataName}{Data imported by ?import.txt.file()}

\item{includeIDs}{IDs of the subjects included for analyses. This must match IDs used in the imported data file.If left empty, IDs will be extracted from the first to the last}

\item{excludeIDs}{IDs to be excluded from the data imported by ?import.txt.file()}

\item{removeMultiallelicSites}{True=remove multi-allelic sites; FALSE=keep multiallelic sites}

\item{remove.nocall.pct}{Variants with percentage of no calls larger than this value will be removed.}

\item{QD.threshold}{Variants with QD smaller than this value will be removed.}

\item{saveRData}{TRUE=Save the returned data frame into an RData file.}

\item{savePath}{Specify save path. If not specified the data will be saved at the working directory.}

\item{label}{Specify the label of the saved file.}
}
\value{
A data frame of cleaned annotated data.
}
\description{
This function prepare the data imported from ?import.txt.file for variant filtering.
}
