% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.variant.R
\name{filter.variant}
\alias{filter.variant}
\title{Variant filtering}
\usage{
\method{filter}{variant}(
  data.name,
  affected.id = NULL,
  unaffected.id = NULL,
  inheritance.pattern = NULL,
  femaleIDs = NULL,
  maleIDs = NULL,
  gene.name = NULL,
  frequency.col = NULL,
  freq.threshold = 1,
  protein.altering = FALSE,
  include.UTRs = FALSE,
  include.synonymous = FALSE,
  include.nonframeshift = TRUE,
  CADD.threshold = 0,
  save.path = NULL,
  save.txt = FALSE,
  label = "test",
  save.genelist = FALSE
)
}
\arguments{
\item{data.name}{Data cleaned by ?clean.raw.data()}

\item{affected.id}{IDs of the affected subjects,which should match the IDs in the data file (column name)}

\item{unaffected.id}{IDs of the unaffected subjects,which should match the IDs in the data file (column name)}

\item{inheritance.pattern}{Model of inheritance pattern. #"AR" = autosomal recessive/"AD"=autosomal dominant/"XLD"=X-linked dominant/"XLR"=X-linked recessive. You can specify "AD" for de novo variant filtering. If XLR is specified, female and male IDs should also be provided.}

\item{femaleIDs}{Female and male IDs should also be provided.}

\item{maleIDs}{Female and male IDs should also be provided.}

\item{gene.name}{Specify the gene list to run candidate gene analysis.Example:gene.name = c("ALPL", "COL1A1", "PLS3")}

\item{frequency.col}{Specify the column names of the frequencies to be used for filtering. For example: frequency.col = c("ExAC_All", "X1000G_ALL", "ExAC_NFE", "X1000G_EUR")}

\item{freq.threshold}{The threshold of the frequency used for filtering.}

\item{protein.altering}{Only include protein altering variants.}

\item{include.UTRs}{On top of protein altering variants, TRUE=also include UTRs}

\item{include.synonymous}{On top of protein altering variants, TRUE=also to include synonymous variants}

\item{include.nonframeshift}{TRUE=also to include nonframeshift variants}

\item{CADD.threshold}{The threshold of the CADD score used for filtering}

\item{save.path}{savePath should specified if saveRData is TRUE;otherwise the data will be saved at the working directory.}

\item{save.txt}{TRUE=Save the returned data frame into a txt file.}

\item{label}{Specify the label of the saved file.}

\item{save.genelist}{TRUE=Save the filtered gene list into a txt file.}
}
\value{
A data frame of variants after filtering.
}
\description{
This function filters the variants based on specified criteria and will print the filtering log on the console.
}
