% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTXT.R
\name{importTXT}
\alias{importTXT}
\title{Import txt file from ANNOVAR annotation}
\usage{
importTXT(filePath = NULL)
}
\arguments{
\item{filePath}{file path of the txt file from ANNOVAR annotation}
}
\value{
A data frame of ANNOVAR annotation data.
}
\description{
This function loads the txt file from ANNOVAR annotation as a data frame.The structure of the data will be printed at the console automatically. Please if numeric variables were imported correctly.
}
