% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pedigreeGT.R
\name{plot.pedigreeGT}
\alias{plot.pedigreeGT}
\title{Generate pedigree plots}
\usage{
\method{plot}{pedigreeGT}(
  variant.list,
  pedigree.data,
  includeIDs,
  gene.name = NULL,
  save.path = NULL,
  frequency.col = c("ExAC_All", "X1000G_ALL"),
  label = "test"
)
}
\arguments{
\item{variant.list}{Filtered variant list from ?filter.variant()}

\item{pedigree.data}{Pedigree data load from excel sheet. check ?kinship2::pedigree() for correct format.}

\item{includeIDs}{IDs of the subjects included for analyses. This must match IDs used in the imported data file.If left empty, IDs will be extracted from the first to the last}

\item{gene.name}{specify the list of genes in which the variants are to be visualized in the pedigree plots}

\item{save.path}{Specify save path. If not specified the data will be saved at the working directory.}

\item{frequency.col}{Column names of the frequencies to be visualized in pedigree plot. For example: frequency.col = c("ExAC_All", "X1000G_ALL", "ExAC_NFE", "X1000G_EUR")}

\item{label}{Specify the label of the folder to save the files.}
}
\value{
Pedigree plot saved as PNG files.
}
\description{
This function generates pedigree plots to visualize variant information and individual genotypes.
}
